[IMP] stock_weighing: quantity_done
Until now, the weighing cards just showed the recorded quantity, which
is the one the user set from the weighing wizard. This could lead to
confussion when a move has a quantity set from the wizard.

For the sake of clarity we put the quantity done while the logic for the
recorded_weight field stays the same and could be shown in the future as
a complementary information.

TT52976