[FIX] stock_weighing_remote_measure: stop weighing when adding a tare
If the weighing continues the tare can be not added to the weight, it seems to be caused by a race condition, but I haven’t been able to prevent it.

To avoid the problem right now, the only way that i found to do it is stopping the measure before registering a tare.